remove old wind estimator
