+++ src/main/flight/ahrs.h
@@ -55,7 +55,12 @@ void ahrsUpdate(timeUs_t currentTimeUs);
 bool ahrsIsHealthy(void);
 bool isAhrsHeadingValid(void);
 float ahrsGetTiltAngle(void);
+
 void updateWindEstimator(void);
+// Wind velocity vectors in cm/s relative to the earth frame
+float getEstimatedWindSpeed(int axis);
+// Returns the horizontal wind velocity as a magnitude in cm/s and, optionally, its heading in EF in 0.01deg ([0, 360 * 100)).
+float getEstimatedHorizontalWindSpeed(uint16_t *angle);
 
 void ahrsTransformVectorBodyToEarth(fpVector3_t * v);
 void ahrsTransformVectorEarthToBody(fpVector3_t * v);
