+++ src/main/flight/wind_estimator.h
@@ -25,13 +25,6 @@
 #include "common/axis.h"
 #include "common/time.h"
 
-bool isEstimatedWindSpeedValid(void);
-// wind velocity vectors in cm / sec relative to the earth frame
-float getEstimatedWindSpeed(int axis);
-// Returns the horizontal wind velocity as a magnitude in cm/s and,
-// optionally, its heading in EF in 0.01deg ([0, 360*100)).
-float getEstimatedHorizontalWindSpeed(uint16_t *angle);
 
-//void updateWindEstimator(timeUs_t currentTimeUs);
 
 #endif
